backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
thefuzz>=0.22.1
python-levenshtein>=0.25.1
pandas>=2.2.2
pydantic>=2.7.1
